Split Binary / JPA class.
