+++ MANO-API/src/main/java/com/ubiqube/etsi/mano/repository/jpa/AbstractJpa.java
@@ -1,8 +1,5 @@
 package com.ubiqube.etsi.mano.repository.jpa;
 
-import java.io.ByteArrayInputStream;
-import java.io.IOException;
-import java.io.InputStream;
 import java.util.List;
 import java.util.Map;
 import java.util.Optional;
@@ -15,17 +12,11 @@ import javax.persistence.criteria.CriteriaQuery;
 import javax.persistence.criteria.From;
 import javax.persistence.criteria.Predicate;
 import javax.persistence.criteria.Root;
-import javax.validation.constraints.NotNull;
 
-import org.apache.commons.io.IOUtils;
-
-import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.ubiqube.etsi.mano.dao.mano.VnfPackage;
-import com.ubiqube.etsi.mano.exception.GenericException;
 import com.ubiqube.etsi.mano.exception.NotFoundException;
 import com.ubiqube.etsi.mano.grammar.AstBuilder;
-import com.ubiqube.etsi.mano.repository.BinaryRepository;
 import com.ubiqube.etsi.mano.repository.ContentManager;
 import com.ubiqube.etsi.mano.repository.CrudRepository;
 
@@ -38,21 +29,18 @@ import ma.glasnost.orika.MapperFacade;
  * @param <T> T is the Json object class.
  * @param <U> Is the Database class.
  */
-public abstract class AbstractJpa<T, U> implements CrudRepository<T>, BinaryRepository {
+public abstract class AbstractJpa<T, U> extends BinaryRepositoryImpl implements CrudRepository<T> {
 	private final EntityManager em;
 	private final org.springframework.data.repository.CrudRepository<U, UUID> repository;
 	private final MapperFacade mapper;
-	private final ContentManager contentManager;
-	private final ObjectMapper jsonMapper;
+
 	private final JpaQueryer queryer;
 
 	public AbstractJpa(final EntityManager em, final org.springframework.data.repository.CrudRepository<U, UUID> repository, final MapperFacade mapper, final ContentManager contentManager, final ObjectMapper jsonMapper) {
-		super();
+		super(contentManager, jsonMapper);
 		this.em = em;
 		this.repository = repository;
 		this.mapper = mapper;
-		this.contentManager = contentManager;
-		this.jsonMapper = jsonMapper;
 		queryer = new JpaQueryer(em);
 	}
 
@@ -98,45 +86,5 @@ public abstract class AbstractJpa<T, U> implements CrudRepository<T>, BinaryRepo
 	}
 
 	abstract Map<String, From<?, ?>> getJoin(Root<U> root);
-
-	@Override
-	public final void storeObject(final String _id, final String _filename, final Object _object) {
-		try {
-			final String str = jsonMapper.writeValueAsString(_object);
-			storeBinary(_id, _filename, new ByteArrayInputStream(str.getBytes()));
-		} catch (final JsonProcessingException e) {
-			throw new GenericException(e);
-		}
-	}
-
-	@Override
-	public final void storeBinary(final String _id, final String _filename, final InputStream _stream) {
-		contentManager.store(_id, _filename, _stream);
-	}
-
-	@Override
-	public final byte[] getBinary(final String _id, final String _filename) {
-		return getBinary(_id, _filename, 0, null);
-	}
-
-	@Override
-	public final byte[] getBinary(final String _id, final String _filename, final int min, final Integer max) {
-		final InputStream os = contentManager.load(_id, _filename, 0, null);
-		try {
-			return IOUtils.toByteArray(os);
-		} catch (final IOException e) {
-			throw new GenericException(e);
-		}
-	}
-
-	@Override
-	public final <T, U extends Class> T loadObject(@NotNull final String _id, final String _filename, final U t) {
-		final byte[] content = getBinary(_id, _filename, 0, null);
-		try {
-			return (T) jsonMapper.readValue(content, t);
-		} catch (final IOException e) {
-			throw new GenericException(e);
-		}
-	}
 
 }
