+++ MANO-API/src/main/resources/application.properties
@@ -19,3 +19,5 @@ spring.artemis.port=61616
 spring.artemis.user=artemis
 spring.artemis.password=simetraehcapa
 jms.queue.destination=demoQueue
+spring.datasource.url = jdbc:h2:file:~/testdb
+spring.jpa.hibernate.ddl-auto = update
